frontend add money done

+++ packages/api/src/bezoekers/bezoekers.resolver.ts
@@ -1,4 +1,4 @@
-import { Resolver, Query, Mutation, Args, Int } from '@nestjs/graphql';
+import { Resolver, Query, Mutation, Args, Int, Float } from '@nestjs/graphql';
 import { BezoekersService } from './bezoekers.service';
 import { Bezoeker } from './entities/bezoeker.entity';
 import { CreateBezoekerInput } from './dto/create-bezoeker.input';
@@ -46,7 +46,7 @@ export class BezoekersResolver {
   }
 
   @Mutation(() => Bezoeker)
-  addSaldoToBezoeker(@Args('uid', { type: () => String }) uid: string, @Args('saldo', { type: () => Int }) aantal: number) {
+  addSaldoToBezoeker(@Args('uid', { type: () => String }) uid: string, @Args('saldo', { type: () => Float }) aantal: number) {
     return this.bezoekersService.addSaldo(uid, aantal);
   }
